+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 Changelog
 ==========
 
+unreleased
+------------------------
+* add flag skip_failed_analyses in CompetitionDay.analyze_flights()
+
 v0.1.0: initial release
 ------------------------
 * competition module: CompetitionDay, Competitor, SoaringSpotDaily, StreplaDaily